fix possible segfault with callbacks
callbacks were creating their userdata as a new cdata of int[1] and then putting an index into that cdata; if the callback wasn't stashed somewhere, this cdata would get released and then the C/native side of things would have a void* to a released cdata.

Instead, now callback userdata is created by directly casting the int index into a void*, so an intermediate int[1] is no longer created.